requirements.txt
# python version ==3.7
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
scikit-image
attrs
h
keras==2.0.8
matplotlib
numpy
pillow
protobuf>=3.0
pydot-ng
scipy
semver
gym[atari]
tensorflow-gpu==1.14
tensorflow==1.14.0
tqdm
moviepy